requirements.txt
streamlit
openpyxl
gspread
oauth2client
Pillow
openai
